+++ extensions/gif-search/src/hooks/useGiphyAPI.ts
@@ -1,6 +1,7 @@
 import { AbortError, FetchError } from "node-fetch";
 import { useCallback, useEffect, useRef, useState } from "react";
 
+import { environment } from "@raycast/api";
 import { GiphyFetch } from "@giphy/js-fetch-api";
 import type { GifsResult } from "@giphy/js-fetch-api";
 import type { IGif as GiphyGif } from "@giphy/js-types";
@@ -77,6 +78,7 @@ export function mapGiphyResponse(giphyResp: GiphyGif) {
     slug: giphyResp.slug,
     preview_gif_url: giphyResp.images.preview_gif.url,
     gif_url: giphyResp.images.fixed_height.url,
-    attribution: "poweredby_giphy.png",
+    attribution:
+      environment.theme === "light" ? "Poweredby_100px-White_VertLogo.png" : "Poweredby_100px-Black_VertLogo.png",
   };
 }